Separate API category pages and remove hero on documentation pages - #1157
