+++ public/index.html
@@ -19,6 +19,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link href="https://fonts.googleapis.com/css?family=Bitter" rel="stylesheet">
     <title>VA API Platform</title>
 
     <!-- We participate in the US government’s analytics program. See the data at analytics.usa.gov. -->
