+++ src/base.scss
@@ -1,5 +1,6 @@
 // Variables
 $font-path: "~uswds/src/fonts";
+$font-bitter: 'Bitter', serif;
 $image-path: "~uswds/src/img";
 $color-gray-cool-light:  #dce4ef;
 $color-primary-alt-light: #9bdaf1;
