+++ src/components/ApiCard.scss
@@ -1,0 +1,34 @@
+a.va-api-card {
+  width: 200px;
+  min-height: 100px;
+  margin: 10px 30px 10px 0;
+  border-top: 5px solid #C4C4C4;
+  text-decoration: none;
+  color: #323A45;
+
+  &:hover {
+    border-top: 5px solid #FDB81E;
+  }
+
+  &:visited {
+    color: #323A45;
+  }
+}
+
+h3.va-api-name {
+  margin: 15px 0 !important;
+  font-size: 2rem;
+}
+
+.va-internal-tag {
+  margin-bottom: 10px;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  font-size: 1.4rem;
+
+  span {
+    padding: 3px 5px;
+    background-color: #DBDBDD;
+    border-radius: 2px;
+  }
+}
