+++ tslint.json
@@ -17,9 +17,9 @@
       },
       "severity": "error"
     },
-    "no-console": {
-      "severity": "error"
-    },
+    // "no-console": {
+    //   "severity": "error"
+    // },
     "prefer-const": {
       "severity": "error"
     },
